requirements.txt
pystray
keyboard
pillow
pythonnet==3.1.0rc0
pywebview
win10toast
setuptools
